refactor(desktop): 抽离 settings 共享类型与 SettingsRow
将 SettingsFormState、SettingsViewProps、常量、formatter 与 SettingsRow 迁入 components/settings/，为后续 panel 拆分做准备。
