+++ apps/desktop/src/components/settings/settings-row.tsx
@@ -1,0 +1,39 @@
+import type { ReactNode } from "react";
+
+import { Label } from "@/components/ui/label";
+import { cn } from "@/lib/utils";
+
+export function SettingsRow({
+  label,
+  description,
+  htmlFor,
+  children,
+  className,
+}: {
+  label: string;
+  description?: string;
+  htmlFor?: string;
+  children: ReactNode;
+  className?: string;
+}) {
+  return (
+    <div
+      className={cn(
+        "flex flex-col gap-3 border-b border-border/35 py-4 last:border-b-0 sm:flex-row sm:items-center sm:justify-between sm:gap-6",
+        className,
+      )}
+    >
+      <div className="min-w-0 sm:max-w-[42%]">
+        <Label htmlFor={htmlFor} className="text-sm font-medium text-foreground">
+          {label}
+        </Label>
+        {description ? (
+          <p className="mt-1 text-xs leading-relaxed text-muted-foreground">{description}</p>
+        ) : null}
+      </div>
+      <div className="min-w-0 w-full sm:w-auto sm:max-w-[min(24rem,52vw)] sm:flex-1 sm:flex sm:justify-end">
+        {children}
+      </div>
+    </div>
+  );
+}
